Add files via upload
The python library uses multinomialNB classification to classify the email as potential or non-potential. The classifier is modelled based on training data devised by the development team, resulting in supervised learning. The structured information contain the source and destination of travelling, date and time, phone number, booked ticket reference number